Phase 1a review: reorg TBAAPI helpers + drop top-of-file comments
- Move `APIEvent+Helpers.swift` and `WeekEventsGrouping.swift` from
  `Services/` to `Extensions/TBAAPI/` — they're model/extension code, not
  services.
- Split the old `Services/TBAAPI+App.swift` into
  `Extensions/TBAAPI/TBAAPIError.swift` (the app-facing error type) and
  `Extensions/TBAAPI/TBAAPI+Events.swift` (the `eventsByYear` wrapper).
  Gives each new domain a clean home for its wrapper extension.
- Drop file-top / class-top narrative blocks on `APIEvent+Helpers`,
  `WeekEventsGrouping`, `EventsListViewController`, and the
  `WeekEventsDelegate` protocol. Inline comments stay.
