+++ the-blue-alliance-ios/Extensions/TBAAPI/TBAAPI+Events.swift
@@ -1,18 +1,6 @@
 import Foundation
 import TBAAPI
 
-// App-facing error type for TBAAPI calls. Each wrapper translates the
-// generated `Operations.*.Output` enum into either a decoded body or a
-// `TBAAPIError`, so call sites only have to deal with async throws.
-enum TBAAPIError: Error {
-    case notModified
-    case unauthorized
-    case notFound
-    case unexpectedStatus(Int)
-}
-
-// Thin wrappers that hide the generated `Operations.*.Output` plumbing.
-// Each phase of the Core Data removal adds the wrappers it needs here.
 extension TBAAPI {
 
     func eventsByYear(_ year: Int) async throws -> [Components.Schemas.Event] {
