+++ the-blue-alliance-ios/Extensions/TBAAPI/TBAAPIError.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+enum TBAAPIError: Error {
+    case notModified
+    case unauthorized
+    case notFound
+    case unexpectedStatus(Int)
+}
